refactor(api): move prepareTopicExpansion to service layer, fix skipped-topics dedupe
- Move prepareTopicExpansion and normalizeTopicTitle from routes/books.ts
  into services/curriculum.ts to honour the route/service boundary (G1/G5)
- Fix skipped-topics dedupe: seed existingTitleKeys only from non-skipped
  topics so skipped titles do not block re-expansion with corrected content
- Add notes and session-crud integration tests; extend progress pagination
  schema (childSessionsQuerySchema / childSessionsPageResponseSchema)
- Add my-notes archive screens, notes hook, session streaming hardening,
  subject-classification hook, and supporting i18n strings
- Add missing parentView i18n keys to all 6 non-English locales
